Conflicts:
	besch/bruecke_besch.h
	besch/fabrik_besch.h
	besch/reader/factory_reader.cc
	besch/vehikel_besch.cc
	besch/vehikel_besch.h
	besch/way_obj_besch.h
	besch/weg_besch.h
	simcity.cc
	simfab.cc
	simline.cc
	vehicle/simvehikel.cc

+++ dataobj/umgebung.h
@@ -49,7 +49,7 @@ public:
 	// if we are the server, we are at this port ...
 	static uint16 server;
 
-	static uint8 announce_server;
+	static uint32 announce_server;
 	static std::string server_name;
 	static std::string server_comment;
